[IMP] stock_reception_screen: switch the screen to a dedicated data model
