+++ oca_dependencies.txt
@@ -32,5 +32,4 @@ stock-logistics-warehouse https://github.com/guewen/stock-logistics-warehouse wm
 # stock-logistics-warehouse https://github.com/guewen/stock-logistics-warehouse 12.0-add-stock_picking_zone
 # stock-logistics-warehouse https://github.com/guewen/stock-logistics-warehouse 12.0-add-reservation-zone-rules
 
-
-
+web
