+++ stock_reception_screen/models/__init__.py
@@ -1,4 +1,5 @@
 from . import stock_move
 from . import stock_move_line
 from . import stock_picking
+from . import stock_reception_screen
 from . import manual_barcode
