+++ stock_reception_screen/models/manual_barcode.py
@@ -13,7 +13,7 @@ class StockPickingManualBarcode(models.TransientModel):
     def button_save(self):
         self.ensure_one()
         record_id = self.env.context.get('active_id')
-        record = self.env["stock.picking"].browse(record_id).exists()
+        record = self.env["stock.reception.screen"].browse(record_id).exists()
         if not record:
             return
         if self.barcode:
