+++ stock_reception_screen/models/stock_move.py
@@ -1,7 +1,7 @@
 # Copyright 2019 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl)
 
-from odoo import fields, models
+from odoo import models
 
 
 class StockMove(models.Model):
@@ -10,7 +10,7 @@ class StockMove(models.Model):
     def action_select_move(self):
         """Set the move as the current one at the picking level."""
         self.ensure_one()
-        self.picking_id.button_cancel_step()
-        self.picking_id.current_move_id = self
-        self.picking_id.next_step()
+        self.picking_id.reception_screen_id.button_cancel_step()
+        self.picking_id.reception_screen_id.current_move_id = self
+        self.picking_id.reception_screen_id.next_step()
         return True
