+++ stock_reception_screen/models/stock_move_line.py
@@ -22,9 +22,9 @@ class StockMoveLine(models.Model):
     def action_select_move_line(self):
         """Set the move line as the current one at the picking level."""
         self.ensure_one()
-        self.picking_id.button_cancel_step()
-        self.picking_id.current_move_line_id = self
-        self.picking_id.current_move_id = self.move_id
-        self.picking_id.next_step()
+        self.picking_id.reception_screen_id.button_cancel_step()
+        self.picking_id.reception_screen_id.current_move_line_id = self
+        self.picking_id.reception_screen_id.current_move_id = self.move_id
+        self.picking_id.reception_screen_id.next_step()
         # FIXME: don't know how to close the pop-up and refresh the screen
         return self.picking_id.action_reception_screen_open()
