--- stock_reception_screen/readme/ROADMAP.rst
@@ -1,2 +1,0 @@
-* this module is not compatible with the Odoo EE module `stock_barcode`
-  (both inherits `stock.picking` from `barcodes.barcode_events_mixin`).
