+++ stock_reception_screen/security/ir.model.access.csv
@@ -1,0 +1,3 @@
+id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+access_stock_reception_screen_user,stock.reception.screen user,model_stock_reception_screen,stock.group_stock_user,1,1,1,1
+access_stock_reception_screen_manager,stock.reception.screen manager,model_stock_reception_screen,stock.group_stock_manager,1,1,1,1
