+++ stock_reception_screen/views/stock_picking.xml
@@ -32,156 +32,4 @@
     </field>
   </record>
 
-  <record id="stock_picking_view_form_screen" model="ir.ui.view">
-    <field name="name">stock.picking.form.screen</field>
-    <field name="model">stock.picking</field>
-    <field name="arch" type="xml">
-      <form class="o_screen">
-
-        <div class="o_screen_header">
-          <div class="o_screen_header_left o_screen_header_content">
-            <field name="name" readonly="1"/>
-          </div>
-          <div class="o_screen_header_center o_screen_header_content">
-            <group>
-            <field name="partner_id" readonly="1" string="Vendor" options="{'no_open': True}"/>
-            <field name="origin" readonly="1"/>
-            </group>
-          </div>
-          <div class="o_screen_header_right o_screen_header_content">
-            <label for="state"/>
-            <field name="state"/>
-          </div>
-        </div>
-
-        <div class="o_screen_actions">
-          <div class="o_screen_content">
-            <button name="action_reception_screen_manual_barcode" type="object"
-                    class="btn-secondary"
-                    string="" icon="fa-terminal"
-                    aria-label="Barcode Input" title="Barcode Input"
-                    attrs="{'invisible': [('current_step', '=', 'done')]}"/>
-          </div>
-          <div class="o_screen_content o_screen_content_right">
-            <button name="button_save_step"
-                    type="object"
-                    string="Next"
-                    icon="fa-check"
-                    class="btn-primary"
-                    barcode_trigger="save"
-                    attrs="{'invisible': [('current_step', 'in', ('select_move', 'done'))]}"/>
-            <button name="button_cancel_step"
-                    type="object"
-                    string="Cancel"
-                    icon="fa-undo"
-                    class="btn-danger"
-                    barcode_trigger="save"
-                    attrs="{'invisible': [('current_step', 'in', ('select_move', 'done'))]}"/>
-            <button name="action_reception_screen_close" type="object"
-              string="" class="btn-secondary" icon="fa-sign-out"
-              aria-label="Exit" title="Exit"/>
-          </div>
-        </div>
-
-        <div class="o_screen_operation bg-primary jumbotron jumbotron-fluid">
-          <div class="container">
-            <field name="current_step_descr"/>
-          </div>
-        </div>
-
-        <field name="current_step" invisible="1"/>
-        <field name="current_move_id" invisible="1"/>
-        <field name="current_move_line_id" invisible="1"/>
-
-        <div class="o_screen_data">
-          <div class="o_screen_data_content" name="data">
-
-            <group attrs="{'invisible': [('current_move_id', '=', False)]}">
-              <label for="current_move_product_display_name"/>
-              <div>
-                <field name="current_move_product_display_name" class="mr8"/>
-              </div>
-              <hr colspan="2"/>
-            </group>
-            <group attrs="{'invisible': [('current_move_line_id', '=', False)]}" name="step_in_progress">
-              <label for="current_move_line_lot_name"/>
-              <div>
-                <field name="current_move_line_lot_name"
-                  class="mr8"
-                  style="width: 15em;"
-                  attrs="{'readonly': [('current_step', '!=', 'set_lot_number')]}"/>
-              </div>
-              <label for="current_move_line_lot_life_date"/>
-              <div>
-                <field name="current_move_line_lot_life_date"
-                  class="mr8"
-                  style="width: 15em;"
-                  attrs="{'readonly': [('current_step', '!=', 'set_expiry_date')]}"/>
-              </div>
-              <label for="current_move_line_qty_done"/>
-              <div colspan="1">
-                <field name="current_move_line_qty_done"
-                  style="width: 15em;"
-                  attrs="{'readonly': [('current_step', '!=', 'set_quantity')]}"/>
-                <field name="current_move_line_uom_id" options="{'no_open': True}"/>
-                <button name="write" type="object"
-                  string="Check" class="btn-primary"
-                  attrs="{'invisible': [('current_step', '!=', 'set_quantity')]}"
-                  style="padding: 0.4em;"/>
-                <field name="current_move_line_qty_status" invisible="1"/>
-                <div class="o_screen_disk" style="background-color: red;"
-                  attrs="{'invisible': [('current_move_line_qty_status', '!=', 'lt')]}"/>
-                <div class="o_screen_disk" style="background-color: green;"
-                  attrs="{'invisible': [('current_move_line_qty_status', '!=', 'eq')]}"/>
-                <div class="o_screen_disk" style="background-color: dodgerblue;"
-                  attrs="{'invisible': [('current_move_line_qty_status', '!=', 'gt')]}"/>
-              </div>
-              <hr colspan="2"/>
-            </group>
-            <group attrs="{'invisible': [('current_step', 'not in', ('select_move', 'done'))]}">
-              <field name="move_lines" nolabel="1" readonly="1">
-                <tree create="0" edit="0" delete="0" decoration-danger="quantity_done &lt; product_uom_qty" decoration-success="quantity_done == product_uom_qty" decoration-info="quantity_done &gt; product_uom_qty">
-                  <field name="product_id"/>
-                  <field name="date_expected"/>
-                  <field name="product_uom_qty" invisible="1"/>
-                  <field name="quantity_done"/>
-                  <field name="product_uom"/>
-                  <button name="action_select_move" type="object"
-                    string="Select" class="btn-primary" style="padding: 0.2em;"/>
-                  <field name="state" invisible="1"/>
-                </tree>
-                <form>
-                  <group>
-                    <field name="move_line_ids" nolabel="1" readonly="1">
-                      <tree create="0" edit="0" delete="0" decoration-success="qty_done > 0">
-                        <field name="product_id"/>
-                        <field name="lot_name"/>
-                        <field name="lot_life_date"/>
-                        <field name="qty_done"/>
-                        <field name="product_uom_id"/>
-                        <button name="action_select_move_line" type="object"
-                          string="Select" class="btn-primary" style="padding: 0.2em;"/>
-                      </tree>
-                      <form>
-                        <group>
-                          <field name="product_id" options="{'no_open': True}"/>
-                          <field name="lot_name"/>
-                          <field name="lot_life_date"/>
-                          <field name="qty_done"/>
-                          <field name="product_uom_id" options="{'no_open': True}"/>
-                        </group>
-                      </form>
-                    </field>
-                  </group>
-                </form>
-              </field>
-            </group>
-          </div>
-        </div>
-
-        <field name="_barcode_scanned" widget="barcode_handler"/>
-      </form>
-    </field>
-  </record>
-
 </odoo>
